+++ fuzz/fuzz_targets/fuzz_parse_trigram.rs
@@ -2,6 +2,9 @@
 use libfuzzer_sys::fuzz_target;
 
 fuzz_target!(|data: &[u8]| {
-    let _ = ix::varint::decode(data);
-    let _ = ix::trigram::Trigram::from_bytes(data);
+    let mut pos = 0;
+    let _ = ix::varint::decode(data, &mut pos);
+    if data.len() >= 3 {
+        let _ = ix::trigram::from_bytes(data[0], data[1], data[2]);
+    }
 });